后台评价查询订单信息及评价人

+++ business-base/src/main/java/com/fm/business/base/model/evaluation/EvaluationInfo.java
@@ -79,7 +79,7 @@ public class EvaluationInfo extends BaseModel implements Serializable, IAudit {
     /**
      * 评价审核状态
      */
-    public Long status;
+    public Integer status;
     /**
      * 创建时间
      **/
